requirements.txt
ansicon==1.89.0
asgiref==3.2.7
blessed==1.17.4
colorama==0.4.3
Django==3.0.5
django-admin-rangefilter==0.6.1
django-appconf==1.0.4
django-crispy-forms==1.9.0
django-cruds-adminlte==0.0.16
django-cursor-pagination==0.1.4
django-image-cropping==1.4.0
django-jquery==3.1.0
django-material-admin==1.6.7
django-python3-ldap==0.11.3
django-select2==7.2.3
django-tables2==2.3.1
djangoajax==3.2
easy-thumbnails==2.7
jinxed==1.0.0
ldap3==2.7
mysqlclient==1.4.6
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
psycopg2==2.8.5
pyasn1==0.4.8
PyMySQL==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
typesentry==0.2.7
wcwidth==0.1.9
